requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cmake==3.18.4.post1
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
idna==2.10
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
